Refactor wheel functions into base env interface
